refactor: move coverage gap contracts to output crate
